# openapi/openapi.yaml
openapi: 3.1.1
info:
  title: oneSCM Handbook Studio API
  version: 0.12.0
  description: |
    Revisionssichere Erstellung rollen- und spartenspezifischer Benutzerhandbücher aus Markdown-Quellen.
    Jede Operation referenziert über `x-requirements` die zugehörigen User Stories (US-020).
    Operationen mit `x-extension: true` ergänzen die verbindliche Operationsliste aus Masterprompt §10 (siehe docs/01-gap-analyse.md, Lücke 10).
    Fehler werden als `application/problem+json` (RFC 9457) geliefert.
    Authentifizierung (ENTSCHEIDUNG E-15), je nach Betriebsart (`GET /auth/config`):
    - `oidc`: `Authorization: Bearer <JWT>` eines OpenID-Connect-Providers; technische Berechtigungen aus einem Gruppen-/Rollen-Claim.
    - `demo`: Header `X-User-Id` mit einem Demo-Benutzer; ohne Header nur Lesezugriff. Nur für Entwicklung und Demo.
    Projekt (ADR-014): Header `X-Project-Id` wählt das Projekt aller projektbezogenen Endpunkte (Standard `p_default`);
    IDs anderer Projekte werden wie nicht vorhandene behandelt (404). Ausnahmen: `/projects…` und die Health-Checks.
    Betrieb (ADR-015, ADR-027): jede Antwort trägt `X-Request-Id` (bei aktivem Tracing zusätzlich `X-Trace-Id`, W3C `traceparent` wird übernommen);
    Rate-Limits je Benutzer und Minute (429 mit `Retry-After`), mit `RATE_LIMIT_STORE=db` gemeinsam über alle Instanzen;
    Prometheus-Metriken unter `/metrics` (außerhalb von `/api/v1`, Bearer `METRICS_TOKEN` oder Berechtigung admin).
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-Token des konfigurierten OIDC-Providers (AUTH_MODE=oidc)
    demoUser:
      type: apiKey
      in: header
      name: X-User-Id
  parameters:
...
    ChapterId: { name: chapterId, in: path, required: true, schema: { type: string } }
    VersionId: { name: versionId, in: path, required: true, schema: { type: string } }
    BlockId: { name: blockId, in: path, required: true, schema: { type: string } }
...
  responses:
...
    Unauthorized:
      description: Nicht angemeldet oder Token ungültig
      headers:
        WWW-Authenticate: { schema: { type: string, example: Bearer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem:
      description: Fehler (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        gate: { $ref: '#/components/schemas/GateResult' }
        blockers: { type: array, items: { type: object } }
    RoleCode: { type: string, enum: [dealer, market, mo, hq, all] }
    DivisionCode: { type: string, enum: [car, van, truck, bus, all, unconfirmed] }
...
    ContentBlock:
      type: object
      properties:
        id: { type: string }
        lineageId: { type: string }
        section: { type: string }
        position: { type: integer }
        kind: { type: string, enum: [paragraph, list, note, tip, warning, xref, gap, table, code] }
        text: { type: string }
        mode: { type: string, enum: [generated, manually_edited, locked, needs_regeneration, ai_rewritten, approved] }
        roles: { type: array, items: { $ref: '#/components/schemas/RoleCode' } }
        divisions: { type: array, items: { $ref: '#/components/schemas/DivisionCode' } }
        market: { type: [string, 'null'] }
        release: { type: [string, 'null'] }
        scopeStatus: { type: string, enum: [confirmed, general, unconfirmed] }
        justification: { type: [string, 'null'] }
        comment: { type: [string, 'null'] }
        versionNo: { type: integer }
        sources: { type: array, items: { type: object } }
        sentences:
          description: Satz-Evidenz eines übernommenen KI-Vorschlags (ADR-013); null, wenn nicht KI-umformuliert
          type: [array, 'null']
          items: { type: object, properties: { text: { type: string }, sourceIds: { type: array, items: { type: string } } } }
    BlockPatch:
      type: object
      properties:
        text: { type: string }
        section: { type: string }
        position: { type: integer }
        kind: { type: string }
        roles: { type: array, items: { $ref: '#/components/schemas/RoleCode' } }
        divisions: { type: array, items: { $ref: '#/components/schemas/DivisionCode' } }
        market: { type: [string, 'null'] }
        release: { type: [string, 'null'] }
        scopeStatus: { type: string, enum: [confirmed, general, unconfirmed] }
        justification: { type: [string, 'null'] }
        comment: { type: [string, 'null'] }
        mode: { type: string, enum: [locked, manually_edited] }
        sourceIds: { type: array, items: { type: string } }
        reason: { type: string }
        expectedVersionNo: { type: integer, description: Optimistische Sperre }
    ChapterVersion:
      type: object
      properties:
        id: { type: string }
        chapterId: { type: string }
        versionNo: { type: integer }
        status: { type: string, enum: [draft, in_review, approved, superseded] }
        submittedBy: { type: [string, 'null'] }
        submittedAt: { type: [string, 'null'] }
        submitComment: { type: [string, 'null'] }
        sections:
          type: array
          items:
            type: object
            properties:
              code: { type: string }
              title: { type: string }
              blocks: { type: array, items: { $ref: '#/components/schemas/ContentBlock' } }
        approvals: { type: array, items: { type: object } }
    GateResult:
      type: object
      properties:
        passed: { type: boolean }
        checks:
          type: array
          items:
            type: object
            properties:
              code: { type: string }
              label: { type: string }
              passed: { type: boolean }
              details: { type: array, items: { type: string } }
...
security:
  - oidcBearer: []
  - demoUser: []
paths:
...
  /chapters:
    get:
      operationId: listChapters
      summary: Kapitel mit Unterkapiteln, Abdeckung, Befund- und Versionsstatus
      tags: [Kapitel]
      x-requirements: [US-002, US-010, US-013, NFR-21]
      x-extension: true
      parameters:
        - { name: outline, in: query, schema: { type: string }, description: 'ohne: Kapitel der Quellen; Gliederungsfamilie: Kapitel dieser Handbuch-Variante; all: beide (ADR-034)' }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Kapitel }
  /chapters/{chapterId}:
    get:
      operationId: getChapter
      summary: Kapitelübersicht
      tags: [Kapitel]
      x-requirements: [US-002]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/ChapterId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Kapitel }
        '404': { $ref: '#/components/responses/Problem' }
  /chapters/{chapterId}/compare:
    get:
      operationId: compareChapterVersions
      summary: Zwei Kapitelversionen vergleichen (ohne Parameter vorletzte gegen neueste)
      tags: [Kapitelwerkstatt]
      x-requirements: [US-019, US-009]
      x-extension: true
      parameters:
        - { $ref: '#/components/parameters/ChapterId' }
        - { name: from, in: query, schema: { type: string }, description: Ausgangsversion }
        - { name: to, in: query, schema: { type: string }, description: Zielversion }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200':
          description: Unterschiede je Absatz (added, removed, changed, moved, unchanged)
          content:
            application/json:
              schema:
                type: object
                properties:
                  from: { type: object }
                  to: { type: object }
                  summary: '{ type: object, properties: { added: { type: integer }, removed: { type: integer }, changed: { type: integer }, moved: { type: integer }, unchanged: { type: integer } } }'
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        change: { type: string, enum: [added, removed, changed, moved, unchanged] }
                        section: { type: string }
                        fields: { type: array, items: { type: string } }
                        from: { type: [object, 'null'] }
                        to: { type: [object, 'null'] }
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /chapters/{chapterId}/versions:
    get:
      operationId: listChapterVersions
      summary: Versionen eines Kapitels
      tags: [Kapitel]
      x-requirements: [US-009, US-019]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/ChapterId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Versionen }
  /chapters/{chapterId}/generate:
    post:
      operationId: generateChapter
      summary: Kapitelentwurf aus bestätigten Quellen generieren
      tags: [Kapitel]
      x-requirements: [US-006, US-008, US-009, US-012]
      parameters: [{ $ref: '#/components/parameters/ChapterId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '201': { description: 'Neue Kapitelversion, content: { application/json: { schema: { $ref: ''#/components/schemas/ChapterVersion'' } } }' }
        '409': { description: 'Blockiert durch offene Blocker-Befunde, content: { application/problem+json: { schema: { $ref: ''#/components/schemas/Problem'' } } }' }
        '422': { $ref: '#/components/responses/Problem' }
  /chapter-versions/{versionId}:
    get:
      operationId: getChapterVersion
      summary: Kapitelversion mit Abschnitten, Blöcken und Quellen
      tags: [Kapitelwerkstatt]
      x-requirements: [US-008, US-009, US-010, US-011]
      parameters: [{ $ref: '#/components/parameters/VersionId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Kapitelversion, content: { application/json: { schema: { $ref: ''#/components/schemas/ChapterVersion'' } } }' }
        '404': { $ref: '#/components/responses/Problem' }
  /chapter-versions/{versionId}/gate:
    get:
      operationId: getGate
      summary: Qualitätsgate für eine Kapitelversion auswerten
      tags: [Freigabe]
      x-requirements: [US-012]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/VersionId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Gate-Ergebnis, content: { application/json: { schema: { $ref: ''#/components/schemas/GateResult'' } } }' }
  /chapter-versions/{versionId}/content-blocks:
    post:
      operationId: createContentBlock
      summary: Absatz, Hinweis, Tipp oder Warnung hinzufügen
      tags: [Kapitelwerkstatt]
      x-requirements: [US-009]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/VersionId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BlockPatch' } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '201': { description: 'Block, content: { application/json: { schema: { $ref: ''#/components/schemas/ContentBlock'' } } }' }
        '409': { $ref: '#/components/responses/Problem' }
  /approval-workflow:
    get:
      operationId: getApprovalWorkflow
      summary: Freigabeworkflow des Projekts (Stufen, Zuständige, Mindestanzahl, Frist, Vier-Augen-Prinzip); ohne Festlegung einstufig
      tags: [Freigabe]
      x-requirements: [NFR-12, US-016]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: OK }
    put:
      operationId: setApprovalWorkflow
      summary: Freigabeworkflow festlegen (Berechtigung admin); leere Stufenliste stellt die einstufige Freigabe wieder her; gilt für neue Einreichungen
      tags: [Freigabe]
      x-requirements: [NFR-12, US-016]
      x-extension: true
      requestBody:
        content: { application/json: { schema: { type: object, properties: { fourEyes: { type: boolean }, stages: { type: array, maxItems: 6, items: { type: object, required: [name], properties: { name: { type: string }, approvers: { type: array, items: { type: string } }, minApprovals: { type: integer, minimum: 1 }, dueDays: { type: [integer, 'null'], minimum: 1, maximum: 90 } } } } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: OK }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /approvals/pending:
    get:
      operationId: listMyPendingApprovals
      summary: Offene Entscheidungen der angemeldeten Person (aktuelle Stufe, zuständig, nicht eigene Einreichung)
      tags: [Freigabe]
      x-requirements: [NFR-12, US-016]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: OK }
...
  /chapter-versions/{versionId}/approve:
    post:
      operationId: approveChapterVersion
      summary: 'Fachliche Entscheidung in der aktuellen Freigabestufe: zustimmen (letzte Stufe: freigegeben) oder ablehnen (zurück in den Entwurf); Zuständigkeit und Vier-Augen-Prinzip (ADR-025)'
      tags: [Freigabe]
      x-requirements: [US-012, US-016, NFR-12]
      parameters: [{ $ref: '#/components/parameters/VersionId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [comment]
              properties:
                comment: { type: string }
                decision: { type: string, enum: [approved, rejected], default: approved }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Freigegeben, content: { application/json: { schema: { $ref: ''#/components/schemas/ChapterVersion'' } } }' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { description: 'Gate nicht bestanden oder Version nicht im Entwurf, content: { application/problem+json: { schema: { $ref: ''#/components/schemas/Problem'' } } }' }
  /chapter-versions/{versionId}/submit:
    post:
      operationId: submitChapterVersion
      summary: Kapitelversion zur Freigabe einreichen (nur mit bestandenem Qualitätsgate; danach gesperrt)
      tags: [Freigabe]
      x-requirements: [US-012, US-016]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/VersionId' }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { comment: { type: string } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Eingereicht, content: { application/json: { schema: { $ref: ''#/components/schemas/ChapterVersion'' } } }' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { description: 'Gate nicht bestanden oder falscher Status, content: { application/problem+json: { schema: { $ref: ''#/components/schemas/Problem'' } } }' }
  /chapter-versions/{versionId}/withdraw:
    post:
      operationId: withdrawChapterVersion
      summary: Einreichung zurückziehen (Version wird wieder bearbeitbarer Entwurf)
      tags: [Freigabe]
      x-requirements: [US-016]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/VersionId' }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { reason: { type: string } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Zurückgezogen }
        '409': { $ref: '#/components/responses/Problem' }
  /chapter-versions/{versionId}/evidence:
    get:
      operationId: getEvidence
      summary: Evidenz je Absatz – Quellen, Revision, Bestätigung, Auffälligkeiten
      tags: [Kapitelwerkstatt]
      x-requirements: [US-011, US-012]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/VersionId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Evidenzübersicht }
        '404': { $ref: '#/components/responses/Problem' }
...
  /content-blocks/{blockId}:
    patch:
      operationId: patchContentBlock
      summary: Block ändern, verschieben, klassifizieren, sperren oder entsperren (versioniert)
      tags: [Kapitelwerkstatt]
      x-requirements: [US-009, US-010]
      parameters: [{ $ref: '#/components/parameters/BlockId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BlockPatch' } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Block, content: { application/json: { schema: { $ref: ''#/components/schemas/ContentBlock'' } } }' }
        '409': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: deleteContentBlock
      summary: Block löschen (Soft-Delete, versioniert, wiederherstellbar)
      tags: [Kapitelwerkstatt]
      x-requirements: [US-009]
      parameters:
        - { $ref: '#/components/parameters/BlockId' }
        - { name: reason, in: query, schema: { type: string } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '204': { description: Gelöscht }
        '409': { $ref: '#/components/responses/Problem' }
  /content-blocks/{blockId}/versions:
    get:
      operationId: listContentBlockVersions
      summary: Versionshistorie eines Blocks
      tags: [Kapitelwerkstatt]
      x-requirements: [US-009, US-019]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/BlockId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Versionen }
  /content-blocks/{blockId}/restore:
    post:
      operationId: restoreContentBlock
      summary: Frühere Blockversion wiederherstellen
      tags: [Kapitelwerkstatt]
      x-requirements: [US-009]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/BlockId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [versionNo], properties: { versionNo: { type: integer } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Block, content: { application/json: { schema: { $ref: ''#/components/schemas/ContentBlock'' } } }' }
...